Added ability to edit player profiles
By clicking the "Toggle Edit" button on a player's profile, the Name,
Main, and Secondary fields can be edited.  Clicking the button again
will save the entered values, with the exception of the Name field if
another player with the newly entered name already exists (in which case
the name will be set back to the previous name).